zc-ps2/requirements.txt
# Specify Python version
python==3.11.11

# Data Collection 
requests==2.31.0
scrapy==2.12.0

# PDF Processing
unstructured==0.11.0
unstructured_pytesseract==0.3.15
unstructured-inference==0.6.6
pdfminer.six==20240706
opencv-python==4.11.0.86
pdf2image==1.17.0

# Core data science packages
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.6.1

# Transformers and deep learning
transformers==4.39.3
datasets==2.18.0
torch==2.2.1

# Vector search
pgvector==0.2.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.28

# OpenAI client for LLMs
openai==0.28.0  # For interacting with OpenAI models

# Visualisation
# matplotlib==3.10.1
# seaborn==0.12.0
lets-plot==4.6.0

# Utilities
ipdb==0.13.13
python-dotenv==1.0.1 # Load environment variables from .env file
tqdm==4.67.1
ipykernel==6.29.5
ipywidgets==8.1.5
colorlog>=6.7.0  # For colored logging output